requirements.txt
python-telegram-bot[job-queue]==20.7
openai==1.12.0
python-dotenv==1.0.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
